tmp-mlserving-p1720/sparsity_DNN/requirements.txt
numpy==1.26.4
scipy==1.13.1
transformers==4.42.4
datasets==2.20.0
pillow==10.4.0
huggingface-hub==0.23.4
tqdm==4.66.4
scikit-learn==1.5.1
matplotlib==3.9.1
seaborn==0.13.2
torch --index-url https://download.pytorch.org/whl/cu118
torchvision --index-url https://download.pytorch.org/whl/cu118
torchaudio --index-url https://download.pytorch.org/whl/cu118

